+++ network/discovery/enode.go
@@ -16,7 +16,7 @@ import (
 )
 
 // createLocalNode create a new enode.LocalNode instance
-func createLocalNode(privKey *ecdsa.PrivateKey, storagePath string, ipAddr net.IP, udpPort, tcpPort int) (*enode.LocalNode, error) {
+func createLocalNode(privKey *ecdsa.PrivateKey, storagePath string, ipAddr net.IP, udpPort, tcpPort uint16) (*enode.LocalNode, error) {
 	db, err := enode.OpenDB(storagePath)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not open node's peer database")
@@ -27,7 +27,7 @@ func createLocalNode(privKey *ecdsa.PrivateKey, storagePath string, ipAddr net.I
 	localNode.Set(enr.UDP(udpPort))
 	localNode.Set(enr.TCP(tcpPort))
 	localNode.SetFallbackIP(ipAddr)
-	localNode.SetFallbackUDP(udpPort)
+	localNode.SetFallbackUDP(int(udpPort))
 	localNode.Set(enr.WithEntry("ssv", true))
 
 	return localNode, nil
